Changes to console outputs, added 2-step PC export

+++ src/ledger/key.ts
@@ -21,6 +21,9 @@ async function ledgerGetAccount(derivationPath: string, hrp: string): Promise<{
 	const transport = await TransportNodeHid.open(undefined)
 	const avalanche = new AvalancheApp(transport)
 	const pubkaddr = await avalanche.getAddressAndPubKey(derivationPath, false, hrp)
+    if (pubkaddr.errorMessage != 'No errors') {
+        throw Error(`Can not get address from ledger: ${pubkaddr.errorMessage}, code ${pubkaddr.returnCode}`)
+    }
 	return {
         publicKey: pubkaddr.publicKey.toString('hex'),
         address: pubkaddr.address
@@ -28,6 +31,9 @@ async function ledgerGetAccount(derivationPath: string, hrp: string): Promise<{
 }
 
 export async function initContext(derivationPath: string, hrp: string) {
+    if (fs.existsSync('ctx.json')) {
+        throw Error('ctx.json already exists')
+    }
     const { publicKey, address } = await ledgerGetAccount(derivationPath, hrp)
     const data = new Context(publicKey, address, publicKeyToEthereumAddressString(publicKey), hrp)
     fs.writeFileSync('ctx.json', JSON.stringify(data, null, 2))
